# [REVIEW] Change threshold of optimized code for hash partition