@GPUtester 2020-02-10
Can one of the admins verify this patch?