@trevorsm7 2020-02-10
rerun tests